driver: src: Fixed MisraC violations
This patch fixes the MisraC 10.6 10.7 10.4 10.3 10.1 violations

+++ driver/src/core/xaie_core_aie.c
@@ -57,8 +57,8 @@ AieRC _XAie_CoreConfigureDone(XAie_DevInst *DevInst, XAie_LocType Loc,
 	Mask = CoreMod->CoreEvent->DisableEvent.Mask |
 		CoreMod->CoreEvent->DisableEventOccurred.Mask |
 		CoreMod->CoreEvent->EnableEventOccurred.Mask;
-	Value = XAIE_EVENTS_CORE_INSTR_EVENT_2 <<
-		CoreMod->CoreEvent->DisableEvent.Lsb;
+	Value = (u32)(XAIE_EVENTS_CORE_INSTR_EVENT_2 <<
+		CoreMod->CoreEvent->DisableEvent.Lsb);
 	RegAddr = CoreMod->CoreEvent->EnableEventOff +
 		_XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
 
@@ -91,7 +91,7 @@ AieRC _XAie_CoreEnable(XAie_DevInst *DevInst, XAie_LocType Loc,
 	/* Clear the disable event occurred bit */
 	Mask = CoreMod->CoreEvent->DisableEventOccurred.Mask |
 		CoreMod->CoreEvent->EnableEventOccurred.Mask;
-	Value = 1U << CoreMod->CoreEvent->DisableEventOccurred.Lsb;
+	Value = (u32)(1U << CoreMod->CoreEvent->DisableEventOccurred.Lsb);
 	RegAddr = CoreMod->CoreEvent->EnableEventOff +
 		_XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
 
@@ -102,7 +102,7 @@ AieRC _XAie_CoreEnable(XAie_DevInst *DevInst, XAie_LocType Loc,
 
 	/* Enable the core */
 	Mask = CoreMod->CoreCtrl->CtrlEn.Mask;
-	Value = 1U << CoreMod->CoreCtrl->CtrlEn.Lsb;
+	Value = (u32)(1U << CoreMod->CoreCtrl->CtrlEn.Lsb);
 	RegAddr = CoreMod->CoreCtrl->RegOff +
 		_XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
 
@@ -134,7 +134,7 @@ AieRC _XAie_CoreWaitForDone(XAie_DevInst *DevInst, XAie_LocType Loc,
 	u64 EventRegAddr;
 
 	Mask = CoreMod->CoreEvent->DisableEventOccurred.Mask;
-	Value = 1U << CoreMod->CoreEvent->DisableEventOccurred.Lsb;
+	Value = (u32)(1U << CoreMod->CoreEvent->DisableEventOccurred.Lsb);
 	EventRegAddr = CoreMod->CoreEvent->EnableEventOff +
 		_XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
