+++ driver/src/io_backend/ext/xaie_baremetal.c
@@ -239,7 +239,7 @@ static AieRC XAie_BaremetalIO_BlockWrite32(void *IOInst, u64 RegOff,
 		const u32 *Data, u32 Size)
 {
 	for(u32 i = 0U; i < Size; i++) {
-		XAie_BaremetalIO_Write32(IOInst, RegOff + i * 4U, *Data);
+		XAie_BaremetalIO_Write32(IOInst, RegOff + (u64)(i * 4U), *Data);
 		Data++;
 	}
 
@@ -266,7 +266,7 @@ static AieRC XAie_BaremetalIO_BlockSet32(void *IOInst, u64 RegOff, u32 Data,
 		u32 Size)
 {
 	for(u32 i = 0U; i < Size; i++) {
-		XAie_BaremetalIO_Write32(IOInst, RegOff+ i * 4U, Data);
+		XAie_BaremetalIO_Write32(IOInst, RegOff+ (u64)(i * 4U), Data);
 	}
 
 	return XAIE_OK;
@@ -345,7 +345,7 @@ static AieRC XAie_BaremetalMemFree(XAie_MemInst *MemInst)
 *******************************************************************************/
 static AieRC XAie_BaremetalMemSyncForCPU(XAie_MemInst *MemInst)
 {
-	Xil_DCacheInvalidateRange((uintptr_t)MemInst->VAddr,
+	Xil_DCacheInvalidateRange((intptr_t)MemInst->VAddr,
 			(uintptr_t)MemInst->Size);
 
 	return XAIE_OK;
@@ -365,8 +365,8 @@ static AieRC XAie_BaremetalMemSyncForCPU(XAie_MemInst *MemInst)
 *******************************************************************************/
 static AieRC XAie_BaremetalMemSyncForDev(XAie_MemInst *MemInst)
 {
-	Xil_DCacheFlushRange((uintptr_t)MemInst->VAddr,
-			(uintptr_t)MemInst->Size);
+	Xil_DCacheFlushRange((intptr_t)MemInst->VAddr,
+			(intptr_t)MemInst->Size);
 
 	return XAIE_OK;
 }
@@ -456,7 +456,7 @@ static AieRC _XAie_BaremetalIO_NpiMaskPoll(void *IOInst, u64 RegOff, u32 Mask,
 	 * on some profiling, and it may vary between platforms.
 	 */
 	MinTimeOutUs = 200;
-	Count = ((u64)TimeOutUs + MinTimeOutUs - 1U) / MinTimeOutUs;
+	Count = (u32)(((u64)TimeOutUs + MinTimeOutUs - 1U) / MinTimeOutUs);
 
 	while (Count > 0U) {
 		_XAie_BaremetalIO_NpiRead32(IOInst, RegOff, &RegVal);
@@ -518,7 +518,7 @@ static AieRC XAie_BaremetalIO_RunOp(void *IOInst, XAie_DevInst *DevInst,
 				(XAie_ShimDmaBdArgs *)Arg;
 			for(u8 i = 0; i < BdArgs->NumBdWords; i++) {
 				XAie_BaremetalIO_Write32(IOInst,
-						BdArgs->Addr + i * 4U,
+						BdArgs->Addr + (u64)(i * 4U),
 						BdArgs->BdWords[i]);
 			}
 			break;
