+++ driver/src/pm/xaie_reset_aie.h
@@ -19,7 +19,7 @@
 #include "xaiegbl.h"
 
 /************************** Function Prototypes  *****************************/
-AieRC _XAie_RstShims(XAie_DevInst *DevInst, u32 StartCol, u32 NumCols);
+AieRC _XAie_RstShims(XAie_DevInst *DevInst, u8 StartCol, u8 NumCols);
 AieRC _XAie_PmSetPartitionClock(XAie_DevInst *DevInst, u8 Enable);
 u8 _XAie_PmIsTileRequested(XAie_DevInst *DevInst, XAie_LocType Loc);
